Rename src/ -> mephenger/
