+++ mephenger/models/channel.py
