+++ public/views/conversation.kv
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-#:import config src.config.config
+#:import config mephenger.config.config
 
 <MessageLabel@Label>
     size_hint: (1, None)
