+++ public/views/header.kv
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-#:import config src.config.config
+#:import config mephenger.config.config
 
 <Header@BoxLayout>
     orientation: "horizontal"
